[Flink] Support automatic table creation
